v1.0.1.0 update
v1.0.1.0 (1/5/2015)
- Challonge import tool now stores most recent API Key and Subdomain for
the next time the form is opened. (Fizzi)
- Fixed a few scenarios where the tool would crash. (Fizzi)
- Improved network efficiency of the tool. (Fizzi)
- Added Save/Save As functionality.
- Now prompts to save changes before new/load/close, if changes have not
yet been saved.
- Added minimum matches to rank in settings tab.

v1.0.0.2 (11/11/2014)
- Fixed an issue where having a different default date format set
natively would interfere with date parsing.

+++ SkillKeeper/Toolbox.cs
@@ -25,22 +25,22 @@ namespace SkillKeeper
             return result;
         }
 
-        public static void recalcMatches(List<Person> playerList, List<Match> matchList, Double startMu, Double startSigma, Int32 multiplier, UInt16 decay)
+        public static void recalcMatches(List<Person> playerList, List<Match> matchList, Double startMu, Double startSigma, Double multiplier, UInt16 decay)
         {
             recalcMatches(playerList, matchList, startMu, startSigma, multiplier, decay, DateTime.Today, null);
         }
 
-        public static void recalcMatches(List<Person> playerList, List<Match> matchList, Double startMu, Double startSigma, Int32 multiplier, UInt16 decay, ProgressBar progress)
+        public static void recalcMatches(List<Person> playerList, List<Match> matchList, Double startMu, Double startSigma, Double multiplier, UInt16 decay, ProgressBar progress)
         {
             recalcMatches(playerList, matchList, startMu, startSigma, multiplier, decay, DateTime.Today, progress);
         }
 
-        public static void recalcMatches(List<Person> playerList, List<Match> matchList, Double startMu, Double startSigma, Int32 multiplier, UInt16 decay, DateTime lastDate)
+        public static void recalcMatches(List<Person> playerList, List<Match> matchList, Double startMu, Double startSigma, Double multiplier, UInt16 decay, DateTime lastDate)
         {
             recalcMatches(playerList, matchList, startMu, startSigma, multiplier, decay, lastDate, null);
         }
 
-        public static void recalcMatches(List<Person> playerList, List<Match> matchList, Double startMu, Double startSigma, Int32 multiplier, UInt16 decay, DateTime lastDate, ProgressBar progress)
+        public static void recalcMatches(List<Person> playerList, List<Match> matchList, Double startMu, Double startSigma, Double multiplier, UInt16 decay, DateTime lastDate, ProgressBar progress)
         {
             Dictionary<String, Person> playerMap = new Dictionary<string, Person>();
             DateTime latestMatch = DateTime.MinValue;
